+++ angr/knowledge_plugins/__init__.py
@@ -5,4 +5,5 @@ from .comments import Comments
 from .data import Data
 from .indirect_jumps import IndirectJumps
 from .labels import Labels
+from .cfg import CFGManager
 from .plugin import KnowledgeBasePlugin
