+++ angr/knowledge_plugins/cfg/cfg_manager.py
@@ -1,0 +1,11 @@
+
+from ..plugin import KnowledgeBasePlugin
+
+
+class CFGManager(KnowledgeBasePlugin):
+
+    def __init__(self):
+        pass
+
+
+KnowledgeBasePlugin.register_default("cfgs", CFGManager)
